# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Smart ERP & BI System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: User login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List all customers
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: classification
          in: query
          schema:
            type: string
            enum: [new, vip, inactive]
      responses:
        "200":
          description: List of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomerRequest"
      responses:
        "201":
          description: Customer created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get customer profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerDetail"
    put:
      operationId: updateCustomer
      tags: [customers]
      summary: Update customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomerRequest"
      responses:
        "200":
          description: Customer updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    delete:
      operationId: deleteCustomer
      tags: [customers]
      summary: Delete customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /customers/{id}/calls:
    get:
      operationId: getCustomerCalls
      tags: [customers]
      summary: Get call logs for a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Call logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CallLog"
    post:
      operationId: addCustomerCall
      tags: [customers]
      summary: Add call log
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCallLogRequest"
      responses:
        "201":
          description: Call log added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CallLog"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    LoginResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        message:
          type: string
      required: [user]
    User:
      type: object
      properties:
        id:
          type: integer
        companyId:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, procurement, sales, inventory]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, companyId, name, email, role, isActive, createdAt]
...
    Customer:
      type: object
      properties:
        id:
          type: integer
        companyId:
          type: integer
        name:
          type: string
        businessType:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        classification:
          type: string
          enum: [new, vip, inactive]
        totalPurchases:
          type: number
        lastOrderDate:
          type: string
          format: date-time
          nullable: true
        nextFollowupDate:
          type: string
          format: date
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, companyId, name, classification, totalPurchases, createdAt]
    CustomerDetail:
      allOf:
        - $ref: "#/components/schemas/Customer"
        - type: object
          properties:
            topProducts:
              type: array
              items:
                $ref: "#/components/schemas/TopProduct"
            recentInvoices:
              type: array
              items:
                $ref: "#/components/schemas/SalesInvoice"
          required: [topProducts, recentInvoices]
    CreateCustomerRequest:
      type: object
      properties:
        name:
          type: string
        businessType:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        classification:
          type: string
          enum: [new, vip, inactive]
        nextFollowupDate:
          type: string
          format: date
          nullable: true
      required: [name, classification]
    CallLog:
      type: object
      properties:
        id:
          type: integer
        customerId:
          type: integer
        userId:
          type: integer
          nullable: true
        userName:
          type: string
          nullable: true
        summary:
          type: string
        outcome:
          type: string
          enum: [interested, not_interested, quote_requested, other]
        nextFollowupDate:
          type: string
          format: date
          nullable: true
        calledAt:
          type: string
          format: date-time
      required: [id, customerId, summary, outcome, calledAt]
    CreateCallLogRequest:
      type: object
      properties:
        summary:
          type: string
        outcome:
          type: string
          enum: [interested, not_interested, quote_requested, other]
        nextFollowupDate:
          type: string
          format: date
          nullable: true
      required: [summary, outcome]
    SalesInvoice:
      type: object
      properties:
        id:
          type: integer
        companyId:
          type: integer
        invoiceNumber:
          type: string
        customerId:
          type: integer
        customerName:
          type: string
        status:
          type: string
          enum: [draft, confirmed, cancelled]
        totalAmount:
          type: number
        taxPercent:
          type: number
        discountPercent:
          type: number
        netAmount:
          type: number
        notes:
          type: string
          nullable: true
        createdByUserId:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, companyId, invoiceNumber, customerId, customerName, status, totalAmount, taxPercent, discountPercent, netAmount, createdAt]
...
    TopProduct:
      type: object
      properties:
        productId:
          type: integer
        productName:
          type: string
        productCode:
          type: string
        totalQuantity:
          type: number
        totalRevenue:
          type: number
        orderCount:
          type: integer
      required: [productId, productName, productCode, totalQuantity, totalRevenue, orderCount]
...